Fixing typo h_chain to h_offset

+++ antipasti/preprocessing/preprocessing.py
@@ -203,8 +203,8 @@ class Preprocessing(object):
                 h_chain = 'A' 
                 l_chain = 'B'
                 idx_list = [0]
-                h_range = range(26-self.h_chain, hupsymchain-self.h_chain)
-                l_range = range(24-self.l_chain, lupsymchain-self.l_chain)
+                h_range = range(26-self.h_offset, hupsymchain-self.h_offset)
+                l_range = range(24-self.l_offset, lupsymchain-self.l_offset)
                 h_pos = start_chain
                 l_pos = start_chain
                 
@@ -538,10 +538,11 @@ class Preprocessing(object):
         file_name = pdb_id + self.selection
         path = self.test_structure_path + file_name + self.file_type_input
         new_path = self.test_dccm_map_path + pdb_id
-        self.generate_cdr1_to_cdr3_pdb(self.test_structure_path+pdb_id+self.file_type_input, lresidues=lresidues, hupsymchain=hupsymchain, lupsymchain=lupsymchain) 
-        subprocess.call(['/usr/local/bin/RScript '+str(self.scripts_path)+'pdb_to_dccm.r '+str(path)+' '+str(new_path)+' '+str(self.modes)], shell=True, stdout=open(os.devnull, 'wb'))
-        if os.path.exists(path):
-            os.remove(path)
+        if not os.path.exists(new_path):
+            self.generate_cdr1_to_cdr3_pdb(self.test_structure_path+pdb_id+self.file_type_input, lresidues=lresidues, hupsymchain=hupsymchain, lupsymchain=lupsymchain) 
+            subprocess.call(['/usr/local/bin/RScript '+str(self.scripts_path)+'pdb_to_dccm.r '+str(path)+' '+str(new_path)+' '+str(self.modes)], shell=True, stdout=open(os.devnull, 'wb'))
+            if os.path.exists(path):
+                os.remove(path)
 
         # Getting lengths, residues and masking
         file_path = sorted(glob.glob(os.path.join(self.test_dccm_map_path, '*'+pdb_id+'.npy')))
